--- backend/dto/users-courses_model.go
@@ -1,11 +1,0 @@
-package model
-
-import "time"
-
-type User_Course struct {
-	User    User `gorm:"foreignkey:UserId"`
-	UserId  uint
-	Course  Course `gorm:"foreignkey:CourseId"`
-	CouseId uint
-	Date    time.Time // time.Time es un date
-}
