--- backend/dto/users/users_model.go
@@ -1,10 +1,0 @@
-package dto
-
-type LoginRequest struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
-
-type LoginResponse struct {
-	Token string `json:"token"`
-}
